// src/app/Model/inventory.ts
export class Inventory{
    id: number;
	itemName: string="";
	itemCompany: string="";
	itemNumber: string="";
	itemType: string="";
	carBrand: string="";
    carModel: string="";
	itemPuchasePrice: number;
	itemSellingPrice: number;
	itemQty: number;
	rackNo: string="";
	itemDesc: string="";
	vendorId: number;
}